requirements.txt
flask
mwapi
pyyaml
toolforge
